files

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator');
 
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
@@ -17,10 +18,28 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
     const errors = validationResult(req);
+
+    if (!image) {
+        return res
+        .status(422)
+        .render('admin/edit-product', {
+            docTitle: "Add Product",
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
+        });
+    }
 
     if (!errors.isEmpty()) {
         return res
@@ -32,7 +51,6 @@ exports.postAddProduct = (req, res, next) => {
                 hasError: true,
                 product: {
                     title: title,
-                    imageUrl: imageUrl,
                     price: price,
                     description: description
                 },
@@ -40,6 +58,8 @@ exports.postAddProduct = (req, res, next) => {
                 validationErrors: errors.array()
             });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
         title: title,
@@ -94,7 +114,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
 
@@ -108,7 +128,6 @@ exports.postEditProduct = (req, res, next) => {
                 hasError: true,
                 product: {
                     title: updatedTitle,
-                    imageUrl: updatedImageUrl,
                     price: updatedPrice,
                     description: updatedDescription,
                     _id: prodId
@@ -127,7 +146,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle
             product.price = updatedPrice
             product.description = updatedDescription
-            product.imageUrl = updatedImageUrl
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path
+            }
             return product
                 .save()
                 .then(result => {
@@ -162,7 +184,14 @@ exports.getAdminProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'))
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
